♻️ Into components
