+++ src/AutoScrum/Pages/Index.razor.cs
@@ -15,24 +15,16 @@ namespace AutoScrum.Pages
 {
     public partial class Index
     {
-        private readonly DailyScrumService _dailyScrum = new();
-
         private const int ContentSpan = 21;
         private const int AnchorSpan = 3;
         private bool IsPageInitializing { get; set; } = true;
-
-        private List<WorkItem>? _cachedWorkItems;
-
         [Inject] public HttpClient HttpClient { get; set; } = null!;
         [Inject] public ConfigService ConfigService { get; set; } = null!;
         [Inject] public AutoMessageService MessageService { get; set; } = null!;
-
-        private MarkupString Output { get; set; } = (MarkupString)"";
-
+        [Inject] public DailyScrumService DailyScrum { get; set; } = null!;
+        [Inject] public OutputService OutputService { get; set; } = null!;
+        
         private AzureDevOpsConnectionInfo? ConnectionInfo { get; set; }
-        private List<User> Users { get; set; } = new();
-        private List<User> IncludedUsers { get; set; } = new();
-
         protected override async Task OnInitializedAsync()
         {
             try
@@ -53,117 +45,10 @@ namespace AutoScrum.Pages
                 return;
             }
             
-            await GetDataFromAzureDevOpsAsync();
+            await DailyScrum.GetDataFromAzureDevOpsAsync(ConnectionInfo);
+            OutputService.Update();
             
             IsPageInitializing = false;
-        }
-
-        private async Task OnConnectionSubmit(AzureDevOpsConnectionInfoRequest connectionInfoRequest) => await GetDataFromAzureDevOpsAsync();
-
-        private AzureDevOpsService GetAzureDevOpsService()
-        {
-            EnsureConnectionInfoValid();
-
-            return new AzureDevOpsService(
-                new AzureDevOpsConfig(ConnectionInfo!.AzureDevOpsOrganization!,
-                    new Uri($"https://{ConnectionInfo.AzureDevOpsOrganization}.visualstudio.com"),
-                    ConnectionInfo.ProjectName!, ConnectionInfo.UserEmail!, ConnectionInfo.PersonalAccessToken!),
-                HttpClient);
-        }
-
-        private async Task<Sprint> GetCurrentSprint(AzureDevOpsService? azureDevOpsService = null)
-        {
-            azureDevOpsService ??= GetAzureDevOpsService();
-            
-            var currentSprint = await azureDevOpsService.GetCurrentSprint();
-            Console.WriteLine("Current Sprint: " + currentSprint.Name);
-
-            return currentSprint;
-        }
-        
-        private async Task GetDataFromAzureDevOpsAsync()
-        {
-            try
-            {
-                var sprint = await GetCurrentSprint();
-
-                _cachedWorkItems = await GetAzureDevOpsService()
-                    .GetWorkItemsForSprint(sprint, ConnectionInfo!.TeamFilterBy);
-
-                Users = GetUniqueUsers(_cachedWorkItems);
-                ReloadUsers();
-
-                ReloadWorkItems();
-
-                StateHasChanged();
-            }
-            catch
-            {
-                MessageService.Error("Critical error while loading data from Azure DevOps");
-                throw;
-            }
-        }
-
-        private void ReloadUsers()
-        {
-            IncludedUsers = Users.Where(x => x.Included).ToList();
-        }
-        
-        private void ReloadWorkItems()
-        {
-            foreach (var item in _cachedWorkItems!)
-            {
-                Console.WriteLine($"{item.Type} {item.Id}: {item.Title}");
-            }
-            
-            _dailyScrum.SetWorkItems(_cachedWorkItems, Users);
-
-            UpdateOutput();
-        }
-
-        private void UpdateOutput()
-        {
-            var markdown = _dailyScrum.GenerateReport(Users.Where(x => x.Included).ToList());
-            var html = Markdig.Markdown.ToHtml(markdown);
-            Output = (MarkupString)html;
-
-            StateHasChanged();
-        }
-
-        private static List<User> GetUniqueUsers(IReadOnlyCollection<WorkItem> workItems)
-        {
-            Dictionary<string, User> users = new();
-
-            if (!workItems.Any())
-            {
-                return users.Values.ToList();
-            }
-            
-            foreach (var wi in workItems.Where(x => !string.IsNullOrWhiteSpace(x.AssignedToEmail) && !string.IsNullOrWhiteSpace(x.AssignedToDisplayName)))
-            {
-                if (!users.ContainsKey(wi.AssignedToEmail!))
-                {
-                    users.Add(wi.AssignedToEmail!, new User(wi.AssignedToDisplayName!, wi.AssignedToEmail!));
-                }
-            }
-
-            return users.Values.ToList();
-        }
-
-        private void UserIncludeChanged(User user)
-        {
-            ReloadUsers();
-            ReloadWorkItems();
-        }
-
-        private void EnsureConnectionInfoValid()
-        {
-            if (ConnectionInfo is null)
-            {
-                throw new ArgumentNullException(nameof(ConnectionInfo));
-            }
-            
-            ConnectionInfo.EnsureValid();
         }
     }
 }
